refactor: memos의 레이아웃 컴포넌트와 페이지 컴포넌트 분리
